requirements.txt
Flask>=2.2.5
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.0
gunicorn==21.2.0
Werkzeug>=2.3.7
Jinja2>=3.0
click>=8.0
itsdangerous>=2.0